check Pro-status, removed links from texts
